First iteration of the receiver code
This code is able to read the Joystick message from the Tx module.
I had issues passing the data around and figured that when sending
the data and receiving it in the read/send rf modules, to use the
size of the structure instead of the variable. It makes sense now
to use sizeof(Structure_name) instead of sizeof(var);.

Another issue was passing variables by reference or by value, and
getting confussed (by default variables are passed by value DOH).

I am missing filing up the motor code to move front/back, left/right
based on the value received.